Rework the night command to stay inside the main dir.
Instead of writing in /tmp, just use the current working directory. This will
make it much easier to see what is built and what the state of the system is,
etc.

I also simplified the situation a bit so that the "archived" builds weren't
available through the combo loader, only the current day is (via symlinking).
At some point I'd like to come back to this and make it so that we can have
custom endpoints for each days' set of builds.